Support Operating System Interface Config
This commit delivers support to configure operating
system interface networking.  It does so using and Anisble
dynamic inventory program that transforms the inventory file
into vars and host vars as well as the host group itself.

An ansible playbook with a jinja template file are then used
to write the /etc/network/interfaces file on each host.

The support for renaming interfaces also exists in the playbook
but is currently disabled due to issues with inconsistent
interface naming.  This will be enhanced in future commits.

Automated unit tests are also added for all of the Python
code.