Change interface naming logic
This change moves the interface naming logic from depending on
the user and config file knowing the original name of the operating
system interfaces for their physical ports and now bases the naming
on the ports MAC addreses.

Separate code will add the MAC address keys for the ports to each node
in the inventory file.