Generic ansible inventory
This change makes the os_interfaces_inventory module which makes ansible
inventory from the genesis inventory file a generic genesis to ansible
inventory translator that is not tied to the create_interfaces playbook.

This opens up the ability to use this dyanmic nventory script  for other
playbooks besides the create interfaces use case.