Reorganize the playbook layout for post-os install
The create_interafces yml was originally written as a standalone
playbook.  This commit moves it to be a list of tasks and adds
additional task lists that can be included.

It introduces a new playbook to drive these other tasks.  This allows
us to keep the task yamls separate so they can be included in
non-OpenStack specific playbooks in the future.

The configure_operating_systems.yml playbook is OpenStack solution
specific since it calls out the controllers host group.