Add data switch MAC address gathering
This commit delivers code which causes the operating systems
to communicate on their data switch ports triggering the switch to
record their MAC addresses.

The code then calls to the Mellanox (data) switch and grabs
the port number to MAC address relationships which it then adds
to the inventory file for all of the nodes listed.