lxc-create has to derive container ipv4 itself
group_vars/all is loaded iniitially and thus we cannot use it to
reference variables that are set during the playbook. With this commit
the container ipv4 address offset is set in two places: lxc-create.yml
and group_vars/all. User should be able to set it in a single place.