Update the configure OS playbooks for key transfer
OpenStack Ansible requires that the public/private keys
be set on the compute nodes and the code is currently only
setting them on the controllers.

This changes the playbook and task layout to allow the keys
to be sent to the controllers and compute nodes while only
putting the inventory file on the controllers.