Don't catch exception when cold resetting bmc
Not sure if this is expected, but when issuing a cold reset to the bmc
pyghmi was generating an ipmi exception. Maybe this is due to the bmc
going offline immediately after the command was issued? For now I'm
removing the error handling and just logging that the cold reset was
issued.