Load localhost host_vars when using dynamic inventory
When running gather_mac_addresses playbook sometimes localhost host_vars
are not loaded. Not sure why is behavior is not consistent. There is no
localhost entry returned the dynamic inventory - possibly it should be
in there? For now any problems can be remendied by calling include_vars
on the localhost task.