set container mgmt ip with single group_var key
In playbooks/group_vars/all there is a key (container_mgmt_ipv4) that
represents the management network ip address assigned to the continer.
The lxc-create.yml playbook has to read in the management IP set in the
config.yml and update the container_mgmt_subnet key, which is also in
the playbooks/group_vars/all file. This variable file is only read at
the beginning of the play so we cannot read the container_mgmt_ipv4 key
directly to set the container mgmt ip (in this initial play it will
alway have a value using the default container_mgmt_subnet). I added a
new key, container_mgmt_ipv4_offset, which can be used both in the
lxc-create.yml play (in conjuction with reading the mgmt subnet from
config.yml) and the container_mgmt_ipv4 key in playbooks/group_vars/all
for later plays.

In summary, now container_mgmt_ipv4 has an container_mgmt_ipv4_offset
key that will set the container management ip address in conjuction with
the ipaddr-mgmt-network key in the config.yml