Add images/preseeds/repos for 14.04.5 & 16.04.1
These images are pulled off the web instead of GSA. Getting all images
now will take a significant amount of time. Soon we should add logic to
only download images that are referenced in config.yml. In the meantime
I would suggest commenting out images that are not needed (and removing
the corresponding preseeds/repo lists).