Add gradle files to enable "PurePower" build.
In Gerrit there is a "PurePower" reviewer that seems to represent an
automated Jenkins build. When the yggdrasil repo was "refreshed" to sync
up with the previous repo the *.gradle files were not included. I had
the build.gradle and check.gradle files locally and am re-adding them
now to allow the build to pass.

Change-Id: I50e22fa3df1626311484001eb31d77cbfd849eeb