feat: Only download OS images that will be used
Added logic to only download os images that are referenced by
cobbler-profile keys in config.yml. Download will be skipped if file is
already present. Also, additional cobbler profiles referencing distros
that are not being used will be skipped.

Note that the license header was removed from *.kopts files. These files
should only contain a single line of options to be passed to the kernel.

Change-Id: I78a835190788386018eb21af3963e4228de701d0