feat: Add more robust dhcp support
Remove hard-coded eth15 from interface template

With the current network templating schema, dhcp networks can be
added to the jinja interface file.  However, the nodes need to
specify <network_name>-addr='' in the inventory file in order to
be added to the hostvars section.  This is an extraneous effort
since, by design, dhcp networks don't have a static address
specified.

The commit changes the dynamic inventory code to look not only
at <network_name>-addr, but also at the network "mode" and
include networks that are definied as dhcp.

With this in mind the eth15 dhcp definition has been removed
from the interfaces_ubuntu.j2 file and the pxe-dhcp network
definition was added to master_inventory.yaml Each node that
referenced the interface file with eth15, now has that network.

The outcome of this change is identical to the previous fix, with
the exception that we can modify the interface information in the
config file now.

Change-Id: I5eed88fe222cf005d6f95cdf68e2f2975a8cc69b