feat: Add users and groups defined in config.yml
Task takes 'user' and 'group' dictionaries defined in config.yml
node-templates and passes options to user and group ansible modules. All
current options are supported and required/default values are as defined
in the module documentation:
http://docs.ansible.com/ansible/user_module.html
http://docs.ansible.com/ansible/group_module.html

No playbook was written, but this task could be called within
configure_operating_systems.yml

No root privilege escalation was used (become/sudo) so it is required to
run this task as remote root user. This is consistent with other post-os
install tasks currently.

Change-Id: I59a09a7d7b5caf6e8a921832869aa80625aa2bfc