feat: Add initial support for veth tunneling
Ulysses has a requirement to allow a specific form of veth tunneling
with bridge networks. ubuntu_interfaces.j2 has been modified to
look for the bridge-port: keyword.  The name assigned to bridge-port
will become the new peer veth, while the bridge veth and the port
specified in eth-port will become the official bridge_ports.

As of this commit, veth is not fully supported in all of it's
capabilities.  This commit solves the use case mentioned above.

Change-Id: Ifffe080fb835546ed6ab75ced7b01c867479c942