feat: Support ip address lists in config networks
Add support of lists of specific IP addresses defined in the 'networks'
dictionary in config.yml. If a network definition has an 'available_ips'
key defined with a list of ip addresses / ip address ranges then those
will be used to assign addresses to the nodes. If 'available_ips' is not
defined then the behavior remains unchanged. The list can contain single
IP addresses or ranges of IP addresses. A range is given with a starting
address and stopping address (inclusive) seperated by a space. e.g.:

  networks:
      external1:
          description: Organization site or external network
          addr: 10.0.16.0/22
          available_ips:
              - 10.0.16.20
              - 10.0.16.35
              - 10.0.16.106 10.0.16.110
              - 10.0.16.115

In the example above the following list of IP addresses would be assigned
to nodes sequentially:

['10.0.16.20',  '10.0.16.35',  '10.0.16.106','10.0.16.107','10.0.16.108',
 '10.0.16.109', '10.0.16.110', '10.0.16.115']

Change-Id: I8848acad2823b06e452b4b7af82d2ff36753d3a3