feat: Add Cluster Software Bootstrap Command
Two new config keys are used to define a script to be run at the end of
genesis:

  software-bootstrap-hosts: (ansible group or host)
  software-bootstrap-cmd: (command to be run)

The command is run through the ansible shell module, which uses the
user's default shell. A shebang line will not change the shell
executable (extra code would be needed to parse and set the module's
optional parameter).

Change-Id: I9da0e108cf90208265c2d2ec4d94ad345155f3c1